# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Constancia Backend API",
    "description": "Backend API contract for the Constancia frontend and Discord bot.",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/v1/auth/magic-link": {
      "post": {
        "operationId": "createMagicLink",
        "summary": "Request a Discord magic link",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "discordUserId": {
                    "type": "string"
                  },
                  "guildId": {
                    "type": "string"
                  }
                },
                "required": ["discordUserId", "guildId"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/verify": {
      "get": {
        "operationId": "verifyMagicLink",
        "summary": "Verify a magic link token",
        "tags": ["auth"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {}
                  },
                  "required": ["status", "data"]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}